+++ model/src/application/component/action_row.rs
@@ -11,6 +11,12 @@ pub struct ActionRow {
     pub components: Vec<Component>,
 }
 
+impl From<ActionRow> for Component {
+    fn from(action_row: ActionRow) -> Self {
+        Self::ActionRow(action_row)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -19,4 +25,6 @@ mod tests {
 
     assert_fields!(ActionRow: components);
     assert_impl_all!(ActionRow: Clone, Debug, Eq, Hash, PartialEq, Send, Sync);
+
+    assert_impl_all!(Component: From<ActionRow>);
 }
